# website/public/openapi/vma.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Votrix Managed Agents",
    "version": "0.1.0",
    "description": "VMA is an API-first agent harness framework that gives every Agent a sandbox for tools, files, and ongoing work.\n\n> **Base URL** `https://api.vma.votrixai.com`\n\nSend your VMA API key in the `x-api-key` header with each request. The key automatically connects the request to the right Organization."
  },
  "paths": {
...
    "/v1/files": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload File",
        "description": "Upload a File using `multipart/form-data`.",
        "operationId": "upload_file_v1_files_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your VMA API key. It identifies the Organization this request can access.",
            "example": "vma_example_key"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_v1_files_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                },
                "example": {
                  "id": "file_1234567890abcdef1234567890abcdef",
                  "type": "file",
                  "filename": "source-material.pdf",
                  "mime_type": "application/pdf",
                  "size_bytes": 184320,
                  "sha256": "90b30e2de3a3d4f11f59bc4863f4d80fba129b5f650f85d9d60b24287d6fef27",
                  "scope": null,
                  "created_at": "2026-08-13T14:30:00Z",
                  "updated_at": "2026-08-13T14:30:00Z"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "files"
        ],
        "summary": "List Files",
        "description": "List Files, optionally limited to one Session with `scope_id`.",
        "operationId": "list_files_v1_files_get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "description": "Only return Files belonging to this scope."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Maximum number of records returned in this page.",
            "example": 20
          },
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "description": "Return records before this resource identifier."
          },
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "description": "Return records after this resource identifier."
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your VMA API key. It identifies the Organization this request can access.",
            "example": "vma_example_key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_FileResponse_"
                },
                "example": {
                  "data": [
                    {
                      "id": "file_1234567890abcdef1234567890abcdef",
                      "type": "file",
                      "filename": "source-material.pdf",
                      "mime_type": "application/pdf",
                      "size_bytes": 184320,
                      "sha256": "90b30e2de3a3d4f11f59bc4863f4d80fba129b5f650f85d9d60b24287d6fef27",
                      "scope": null,
                      "created_at": "2026-08-13T14:30:00Z",
                      "updated_at": "2026-08-13T14:30:00Z"
                    },
                    {
                      "id": "file_2234567890abcdef1234567890abcdef",
                      "type": "file",
                      "filename": "brief.pdf",
                      "mime_type": "application/pdf",
                      "size_bytes": 96214,
                      "sha256": "f5c4a9b8db604b511a4b284b2405584b4cc919bb626a23deed4045b35f2f1c78",
                      "scope": {
                        "type": "session",
                        "id": "sess_1234567890abcdef1234567890abcdef"
                      },
                      "created_at": "2026-08-13T14:30:00Z",
                      "updated_at": "2026-08-13T14:32:00Z"
                    }
                  ],
                  "has_more": false,
                  "first_id": "file_1234567890abcdef1234567890abcdef",
                  "last_id": "file_2234567890abcdef1234567890abcdef"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/files/{file_id}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Retrieve File",
        "description": "Retrieve File metadata without downloading its content.",
        "operationId": "retrieve_file_v1_files__file_id__get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the file addressed by this request.",
            "example": "file_1234567890abcdef1234567890abcdef"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your VMA API key. It identifies the Organization this request can access.",
            "example": "vma_example_key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                },
                "example": {
                  "id": "file_1234567890abcdef1234567890abcdef",
                  "type": "file",
                  "filename": "source-material.pdf",
                  "mime_type": "application/pdf",
                  "size_bytes": 184320,
                  "sha256": "90b30e2de3a3d4f11f59bc4863f4d80fba129b5f650f85d9d60b24287d6fef27",
                  "scope": null,
                  "created_at": "2026-08-13T14:30:00Z",
                  "updated_at": "2026-08-13T14:30:00Z"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "files"
        ],
        "summary": "Delete File",
        "operationId": "delete_file_v1_files__file_id__delete",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the file addressed by this request.",
            "example": "file_1234567890abcdef1234567890abcdef"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your VMA API key. It identifies the Organization this request can access.",
            "example": "vma_example_key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                },
                "example": {
                  "id": "file_1234567890abcdef1234567890abcdef",
                  "deleted": true
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            }
          }
        },
        "description": "Delete File for your Organization."
      }
    },
    "/v1/files/{file_id}/content": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Download File",
        "description": "Get a short-lived download URL for a File.",
        "operationId": "download_file_v1_files__file_id__content_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the file addressed by this request.",
            "example": "file_1234567890abcdef1234567890abcdef"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your VMA API key. It identifies the Organization this request can access.",
            "example": "vma_example_key"
          }
        ],
        "responses": {
          "307": {
            "description": "Redirect to a short-lived URL for the File bytes.",
            "headers": {
              "Location": {
                "description": "Short-lived URL for downloading the File.",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_v1_files_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_v1_files_post"
      },
...
      "DeletedResponse": {
        "properties": {
          "id": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DeletedResponse"
      },
...
      "FileResponse": {
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "file",
            "default": "file"
          },
          "filename": {
            "type": "string"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "size_bytes": {
            "type": "integer"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "filename",
          "size_bytes",
          "created_at",
          "updated_at"
        ],
        "title": "FileResponse"
      },
      "FileScope": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "session",
              "sandbox"
            ],
            "default": "session"
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "FileScope"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListResponse_FileResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FileResponse"
            },
            "type": "array"
          },
          "has_more": {
            "type": "boolean",
            "default": false
          },
          "first_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ListResponse[FileResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array"
          },
          "msg": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "input": {},
          "ctx": {
            "type": "object"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://api.vma.votrixai.com",
      "description": "Votrix Managed Agents (VMA) API"
    }
  ],
...
}